+++ back_end/src/main/java/lk/ijse/gdse/springboot/back_end/service/impl/InventoryServiceImpl.java
@@ -4,6 +4,8 @@ import lk.ijse.gdse.springboot.back_end.dto.InventoryDTO;
 import lk.ijse.gdse.springboot.back_end.entity.Inventory;
 import lk.ijse.gdse.springboot.back_end.repository.InventoryRepo;
 import lk.ijse.gdse.springboot.back_end.service.InventoryService;
+import lk.ijse.gdse.springboot.back_end.service.exception.DuplicateRecordException;
+import lk.ijse.gdse.springboot.back_end.service.exception.NotFoundException;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 
@@ -27,27 +29,44 @@ public class InventoryServiceImpl implements InventoryService {
 
     @Override
     public InventoryDTO getInventoryDetails(String id) {
+        if (!inventoryRepo.existsById(id)){
+            throw new NotFoundException("Item Code does not exists!");
+        }
         return modelMapper.map(inventoryRepo.findById(id).get(), InventoryDTO.class);
     }
 
     @Override
     public void saveInventory(InventoryDTO inventoryDTO) {
+        if (inventoryRepo.existsById(inventoryDTO.getItemCode())){
+            throw new DuplicateRecordException("Item Code is already exists!");
+        }
         inventoryRepo.save(modelMapper.map(inventoryDTO, Inventory.class));
     }
 
     @Override
     public void updateInventory(InventoryDTO inventoryDTO) {
+        if (!inventoryRepo.existsById(inventoryDTO.getItemCode())){
+            throw new NotFoundException("Item Code does not exists!");
+        }
         inventoryRepo.save(modelMapper.map(inventoryDTO, Inventory.class));
     }
 
     @Override
     public void deleteInventory(String id) {
+        if (!inventoryRepo.existsById(id)){
+            throw new NotFoundException("Item Code does not exists!");
+        }
         inventoryRepo.deleteById(id);
     }
 
     @Override
     public List<InventoryDTO> getAllDetailsUseGender(String gender) {
         return inventoryRepo.findByTypeOfGenderContaining(gender).stream().map(inventory -> modelMapper.map(inventory, InventoryDTO.class)).toList();
+    }
+
+    @Override
+    public List<InventoryDTO> getAllDetailsUseCategory(String categary) {
+        return inventoryRepo.findByCategory(categary).stream().map(inventory -> modelMapper.map(inventory, InventoryDTO.class)).toList();
     }
 
     @Override
@@ -58,6 +77,11 @@ public class InventoryServiceImpl implements InventoryService {
     @Override
     public List<InventoryDTO> getAllItemsByPriceAndTypeOfGender(double minPrice, double maxPrice,String gender) {
         return inventoryRepo.findByUnitPriceSaleBetweenAndTypeOfGenderContaining(minPrice,maxPrice,gender).stream().map(inventory -> modelMapper.map(inventory, InventoryDTO.class)).toList();
+    }
+
+    @Override
+    public List<InventoryDTO> searchItemByName(String name) {
+        return inventoryRepo.findByItemDesc(name).stream().map(inventory -> modelMapper.map(inventory,InventoryDTO.class)).toList();
     }
